backend/requirements.txt
# AI Marketing Automation Platform - Backend Dependencies

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database (Python 3.13 compatible versions)
sqlalchemy==1.4.53
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI Services
anthropic==0.40.0

# Social Media APIs
requests==2.31.0
facebook-business==18.0.2

# Image/Video Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Configuration
python-dotenv==1.0.0

# Email
fastapi-mail==1.4.1

# Payment Processing (Indian market)
razorpay==1.3.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.0.0
pytest-timeout==2.1.0
pytest-mock==3.10.0
pytest-xdist==3.0.0
psutil==5.9.0
httpx==0.24.0
factory-boy==3.2.0
faker==18.0.0
black==23.11.0
flake8==6.1.0